# lazy load modules to speed up normal invoke task case